Fixing an error with task options and renaming the config.
